Adding exercise 6
It won't let me re-upload all of the 'cabi_data' in a single folder because it contains more than 100 files. My brain hurts, so instead of trying to fix this myself: could you please ensure that a copy of 'cabi_data' is in the same folder as the rest of my code when running it? Thanks!